Extend the autoloading
