+++ Assets/Scripts/Data/IScriptableObjectGroup.cs
@@ -1,0 +1,11 @@
+/// <summary>
+/// An interface that is used for the <see cref="ScriptableObjectGroup{T}" /> to expose
+/// the methods without exposing the type parameter.
+/// </summary>
+public interface IScriptableObjectGroup
+{
+	/// <summary>
+	/// Loads children objects from the base folder.
+	/// </summary>
+	public void LoadObjects();
+}
