+++ Assets/Scripts/Data/IScriptableObjectGroup.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 83306574f23aff34884e601f755d8248
