+++ Assets/Scripts/Data/ScriptableObjectGroup.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 090b633f123bd9548801fa18a00e6eb5
