--- Assets/Scripts/Editor/Dialogue/DialogueGroupEditor.cs
@@ -1,17 +1,0 @@
-using UnityEditor;
-using UnityEngine;
-
-[CustomEditor(typeof(DialogueGroup))]
-public class DialogueGroupEditor : Editor
-{
-	public override void OnInspectorGUI()
-	{
-		DrawDefaultInspector();
-
-		DialogueGroup group = (DialogueGroup)target;
-		if (GUILayout.Button("Update Nodes"))
-		{
-			group.LoadNodes();
-		}
-	}
-}
