--- Assets/Scripts/Editor/Dialogue/DialogueGroupEditor.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 23002714a82ac5740b0d19216249e419
